Fixed the 'too early death' issue
* was calulating and comparing the tv.sec & tv.usec elements
  in unsigned int variables but it resulted in uint overflows
	- to fix this, I substitute the uint by long long int.
==> look at the mandatory part if there isn't that overflow issue too

+++ philo_bonus/inc/philo_bonus.h
@@ -119,8 +119,8 @@ void			wait_signal(sem_t *signal, unsigned int signals_nb);
 void			set_timers(t_times *timers, char **argv);
 int				is_time_to_stop(t_child *child);
 int				is_time_to_die(t_child *child);
-unsigned int	get_usleep_time(t_times time, sem_t *sem, char action);
-unsigned int	get_thinking_sleep(t_times time);
+long long int	get_usleep_time(t_times time, sem_t *sem, char action);
+long long int	get_thinking_sleep(t_times time);
 
 // utils.c
 unsigned int	ft_atou(char *str_number);
